Fixed the JSON->shell command problems with multiple values for a key
The original implementation already used JSON.loads to load the JSON file
containing the array with objects to be converted to shell commands. There is an
issue with this when there are multiple keys with the same value. In that case
only the last value was retained.

I have now changed the behavior to allow for the inclusion of an array that will
be expended to a repetition of the same key with different arguments. The case
where there is a need to have a single key with an array of arguments is also
supported although in that case the arguments have to be provided in the array
form they would take in the shell command (the shell command is built using
string concatenation).

Lastly this commit also adds documentation to this process within the python
script.